refactor: move PersonalisationBuilder into templates/notify
